refactor: move ConversationGuard::e2ei_rotate() to `commit` module
So we can delete the e2ei module.
