--- crypto/src/mls/conversation/conversation_guard/e2e_identity/mod.rs
@@ -1,4 +1,0 @@
-mod rotate;
-
-use super::ConversationGuard;
-use super::Result;
